Fix conflict resolution after #13 merge
