+++ submodules/FakePasscode/Sources/FakePasscodeSettings.swift
@@ -5,138 +5,6 @@ import TelegramUIPreferences
 import SwiftSignalKit
 import TelegramStringFormatting
 import AccountContext
-import TelegramIntents
-
-public enum PeerRemovalType: Int32, Codable {
-    case delete
-    case hide
-}
-
-public struct PeerWithRemoveOptions: Codable, Equatable {
-    public let peerId: PeerId
-    public let removalType: PeerRemovalType
-    public let deleteFromCompanion: Bool
-    
-    public init(peerId: PeerId, removalType: PeerRemovalType, deleteFromCompanion: Bool) {
-        self.peerId = peerId
-        self.removalType = removalType
-        self.deleteFromCompanion = deleteFromCompanion
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: StringCodingKey.self)
-        self.peerId = try container.decode(PeerId.self, forKey: "pid")
-        self.removalType = PeerRemovalType(rawValue: try container.decode(Int32.self, forKey: "rt")) ?? .delete
-        self.deleteFromCompanion = (try container.decodeIfPresent(Int32.self, forKey: "dfc") ?? 0) != 0
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: StringCodingKey.self)
-        try container.encode(self.peerId, forKey: "pid")
-        try container.encode(self.removalType.rawValue, forKey: "rt")
-        try container.encode((self.deleteFromCompanion ? 1 : 0) as Int32, forKey: "dfc")
-    }
-}
-
-public struct FakePasscodeAccountActionsSettings: Codable, Equatable {
-    public let peerId: PeerId
-    public let recordId: AccountRecordId
-    public let chatsToRemove: [PeerWithRemoveOptions]
-    public let logOut: Bool
-    
-    public static func defaultSettings(peerId: PeerId, recordId: AccountRecordId) -> FakePasscodeAccountActionsSettings {
-        return FakePasscodeAccountActionsSettings(peerId: peerId, recordId: recordId, chatsToRemove: [], logOut: false)
-    }
-    
-    public init(peerId: PeerId, recordId: AccountRecordId, chatsToRemove: [PeerWithRemoveOptions], logOut: Bool) {
-        self.peerId = peerId
-        self.recordId = recordId
-        self.chatsToRemove = chatsToRemove
-        self.logOut = logOut
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: StringCodingKey.self)
-
-        self.peerId = try container.decode(PeerId.self, forKey: "pid")
-        self.recordId = try container.decode(AccountRecordId.self, forKey: "rid")
-        self.chatsToRemove = try container.decodeIfPresent([PeerWithRemoveOptions].self, forKey: "ctr") ?? []
-        self.logOut = (try container.decodeIfPresent(Int32.self, forKey: "lo") ?? 0) != 0
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: StringCodingKey.self)
-
-        try container.encode(self.peerId, forKey: "pid")
-        try container.encode(self.recordId, forKey: "rid")
-        try container.encode(self.chatsToRemove, forKey: "ctr")
-        try container.encode((self.logOut ? 1 : 0) as Int32, forKey: "lo")
-    }
-    
-    public func withUpdatedChatsToRemove(_ chatsToRemove: [PeerWithRemoveOptions]) -> FakePasscodeAccountActionsSettings {
-        return FakePasscodeAccountActionsSettings(peerId: self.peerId, recordId: self.recordId, chatsToRemove: chatsToRemove, logOut: self.logOut)
-    }
-    
-    public func withUpdatedLogOut(_ logOut: Bool) -> FakePasscodeAccountActionsSettings {
-        return FakePasscodeAccountActionsSettings(peerId: self.peerId, recordId: self.recordId, chatsToRemove: self.chatsToRemove, logOut: logOut)
-    }
-    
-    public func performActions(context: AccountContext, updateSettings: @escaping (FakePasscodeAccountActionsSettings) -> Void) {
-        let waitingQueueSizeValue = Atomic<Int>(value: 0)
-        let waitingQueueSizePromise = ValuePromise<Int>(0)
-        
-        let updateWaitingQueueSize: ((Int) -> Int) -> Void = { f in
-            waitingQueueSizePromise.set(waitingQueueSizeValue.modify { f($0) })
-        }
-        
-        for chatToRemove in chatsToRemove {
-            switch chatToRemove.removalType {
-            case .delete:
-                if chatToRemove.peerId.namespace == Namespaces.Peer.CloudChannel {
-                    context.peerChannelMemberCategoriesContextsManager.externallyRemoved(peerId: chatToRemove.peerId, memberId: context.account.peerId)
-                }
-                
-                let deleteGloballyIfPossible: Bool
-                switch chatToRemove.peerId.namespace {
-                case Namespaces.Peer.CloudUser:
-                    deleteGloballyIfPossible = chatToRemove.deleteFromCompanion
-                    break
-                case Namespaces.Peer.SecretChat:
-                    deleteGloballyIfPossible = true
-                    break
-                default:
-                    deleteGloballyIfPossible = false
-                    break
-                }
-                
-                updateWaitingQueueSize({ $0 + 1 })
-                let _ = context.engine.peers.removePeerChat(peerId: chatToRemove.peerId, reportChatSpam: false, deleteGloballyIfPossible: deleteGloballyIfPossible).start(completed: {
-                    deleteSendMessageIntents(peerId: chatToRemove.peerId)
-                    updateWaitingQueueSize({ $0 - 1 })
-                })
-                break
-                
-            case .hide:
-                break
-            }
-        }
-        
-        let updatedChatsToRemove = chatsToRemove.filter({ !($0.peerId.namespace == Namespaces.Peer.SecretChat && $0.removalType == .delete) })
-        if updatedChatsToRemove != chatsToRemove {
-            updateSettings(self.withUpdatedChatsToRemove(updatedChatsToRemove))
-        }
-        
-        if logOut {
-            let _ = (waitingQueueSizePromise.get()
-            |> filter { $0 == 0 }
-            |> take(1)
-            |> deliverOnMainQueue).start(next: { _ in
-                context.sharedContext.applicationBindings.clearAllNotifications()
-                let _ = logoutFromAccount(id: recordId, accountManager: context.sharedContext.accountManager, alreadyLoggedOutRemotely: false).start()
-            })
-        }
-    }
-}
 
 public struct FakePasscodeSmsActionSettings: Codable, Equatable {
     public init() {
@@ -278,9 +146,9 @@ public struct FakePasscodeSettingsHolder: Codable, Equatable {  // TODO probably
         return settings.accountActions.first(where: { $0.peerId == account.peerId && $0.recordId == account.id }) ?? .defaultSettings(peerId: account.peerId, recordId: account.id)
     }
 
-    public func getAccountActions(_ uuid: UUID, _ account: FakePasscodeActionsAccount) -> FakePasscodeAccountActionsSettings {
-        let fakePasscodeSettings = self.settings.first(where: { $0.uuid == uuid })!
-        let settings = fakePasscodeSettings.accountActions.first(where: { $0.peerId == account.peerId && $0.recordId == account.recordId }) ?? .defaultSettings(peerId: account.peerId, recordId: account.recordId)
+    public func getAccountActions(_ uuid: UUID, _ account: Account) -> FakePasscodeAccountActionsSettings {
+        let fakePasscodeSettings = self.settings.first(where: { $0.uuid == uuid })
+        let settings = fakePasscodeSettings?.accountActions.first(where: { $0.peerId == account.peerId && $0.recordId == account.id }) ?? .defaultSettings(peerId: account.peerId, recordId: account.id)
         return settings;
     }
 }
